requirements.txt
Django==1.9
django-allauth==0.24.1
django-crispy-forms==1.6.0
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-watson==1.2.1
dj-database-url==0.2.2
dj-static==0.0.6
gunicorn==18.0
static==0.4
wsgiref==0.1.2
django-registration-redux==1.3
django-toolbelt==0.0.1
Markdown==2.6.6
markdown2==2.3.1
oauthlib==1.0.3
psycopg2==2.6.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.1
virtualenv==13.1.2
whitenoise==3.2

